+++ lib/src/appender/ConsoleAppender.dart
@@ -1,9 +1,16 @@
+import 'dart:async';
 import 'dart:math';
 
+import 'package:intl/intl.dart';
 import 'package:log_4_dart_2/log_4_dart_2.dart';
 import 'package:log_4_dart_2/src/LogRecord.dart';
 import 'package:log_4_dart_2/src/LogRecordFormatter.dart';
 import 'package:log_4_dart_2/src/appender/Appender.dart';
+import 'dart:developer' as devtools;
+
+/// Set devtools this when running on the physical app or on webbrowser. Will log to dart:devtools. This way, the result
+/// will not be a simple string, but the passed in objects (error etc.)
+enum ConsoleLoggerMode { stdout, devtools }
 
 ///
 /// A appender for writing logs to the console output
@@ -11,10 +18,40 @@ import 'package:log_4_dart_2/src/appender/Appender.dart';
 class ConsoleAppender extends Appender {
   static const LOGGER_NAME = 'CONSOLE';
 
+  ConsoleLoggerMode mode = ConsoleLoggerMode.stdout;
+
+  // for devtools logging
+  int sequenceNumber = 1;
+
   @override
   void append(LogRecord logRecord) {
     logRecord.loggerName ??= getType();
-    print(LogRecordFormatter.format(logRecord, format!, dateFormat: dateFormat, brackets: brackets));
+
+    if (mode == ConsoleLoggerMode.devtools) {
+      // void log(
+      //   String message,
+      //   {DateTime? time,
+      //   int? sequenceNumber,
+      //   int level = 0,
+      //   String name = '',
+      //   Zone? zone,
+      //   Object? error,
+      //   StackTrace? stackTrace}
+      // )
+      devtools.log(
+        LogRecordFormatter.eval(logRecord.message),
+        time: logRecord.time,
+        sequenceNumber: sequenceNumber++,
+        level: logRecord.level.value,
+        name: '${logRecord.tag}',
+        zone: Zone.current,
+        error: logRecord.error,
+        stackTrace: logRecord.stackTrace,
+      );
+    } else {
+      print(LogRecordFormatter.format(logRecord, format!, dateFormat: dateFormat, brackets: brackets));
+    }
+
     var tabs = '\t';
     if (logRecord.error != null) {
       print(tabs + logRecord.error.toString());
@@ -58,6 +95,15 @@ class ConsoleAppender extends Appender {
       brackets = config['brackets'];
     } else {
       brackets = false;
+    }
+    if (config.containsKey('mode')) {
+      if (config['mode'] == 'stdout') {
+        mode = ConsoleLoggerMode.stdout;
+      } else if (config['mode'] == 'devtools') {
+        mode = ConsoleLoggerMode.devtools;
+      }
+    } else {
+      mode = ConsoleLoggerMode.stdout;
     }
     return null;
   }