+++ lib/src/constants.dart
@@ -3,7 +3,7 @@ const kStackDepthOfThis = 2;
 // How many wrapping calls are there between client call to library and the Logger.log(...) method?
 // LoggerExposure will here add + 1 to whatever the client already adds. Usually zero, but could be that there are
 // additional filters in the client app that have to be bypassed
-const kStackDepthOffsetForLoggerExposureClass = 1;
+const kStackDepthOffset = 1;
 
 const kOpen = '[';
 const kClose = ']';